+++ tests/test_source_publication.py
@@ -132,7 +132,7 @@ class PublicationAPITest(TestCase):
             source=self.src,
             timeperiod_startdate=["2020-01-01"],
             timeperiod_enddate=["2021-01-01"],
-            provenance={"text_log": "Test provenance"},
+            provenance={},
             status="p"
         )